[S05P31A103-173] fix : 회원탈퇴시 redis 같이 삭제

+++ backend/mbotc/src/main/java/com/ssafy/mbotc/controller/UserController.java
@@ -6,10 +6,19 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.ssafy.mbotc.entity.User;
+import com.ssafy.mbotc.service.RedisService;
 import com.ssafy.mbotc.service.SyncService;
 import com.ssafy.mbotc.service.UserService;
 
@@ -25,7 +34,10 @@ public class UserController {
 	private UserService userService;
 	
 	@Autowired
-	private SyncService syncservice;
+	private SyncService syncService;
+	
+	@Autowired
+	private RedisService redisService;
 	
 	// after login, user token update
 	@PatchMapping
@@ -42,7 +54,7 @@ public class UserController {
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "USER NOT FOUND");
 		}
 		User userResult = userService.updateUserToken(target.get(), user.getToken());
-		syncservice.syncWithUser(userResult.getToken(), userResult.getUrl(), userResult.getUserId());
+		syncService.syncWithUser(userResult.getToken(), userResult.getUrl(), userResult.getUserId());
 		return ResponseEntity.status(HttpStatus.OK).body(userResult);
 	}
 	
@@ -78,7 +90,7 @@ public class UserController {
 		if(user.getToken() != null && user.getToken().equals("cookie")) {
 			user.setToken(mmAuthToken);
 			u = userService.save(user);
-			syncservice.syncWithUser(u.getToken(), u.getUrl(), u.getUserId());
+			syncService.syncWithUser(u.getToken(), u.getUrl(), u.getUserId());
 			return ResponseEntity.status(HttpStatus.OK).body(u);
 		}
 		
@@ -87,12 +99,12 @@ public class UserController {
 			userInfo.get().setToken(user.getToken());
 			userInfo.get().setUserName(user.getUserName());
 			u = userService.save(userInfo.get());
-			syncservice.syncWithUser(u.getToken(), u.getUrl(), u.getUserId());
+			syncService.syncWithUser(u.getToken(), u.getUrl(), u.getUserId());
 			return ResponseEntity.status(HttpStatus.OK).body(u);
 		}
 		
 		u = userService.save(user);		
-		syncservice.syncWithUser(u.getToken(), u.getUrl(), u.getUserId());
+		syncService.syncWithUser(u.getToken(), u.getUrl(), u.getUserId());
 		return ResponseEntity.status(HttpStatus.OK).body(u);
 	}
 	
@@ -117,6 +129,11 @@ public class UserController {
 		if(!target.get().getUserEmail().equals(user.getUserEmail())) {
 			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED");
 		}
+		
+		boolean success = redisService.deleteUserSettings(user.getUserId());
+		if(!success)
+			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
+		
 		userService.delete(target.get());
 		return ResponseEntity.status(HttpStatus.OK).body("SUCCESS");
 	}
